requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
boto3
sqlalchemy
pydantic
pytest
python-multipart
